Consolidated controller data into one js file
